feat: prefer compilation database from current working directory
To support working with conan workspaces where there is
a compile_commands.json linked into the root folder and potentially one
into the individual conan package subfolders, this change prefers the
compile_commands.json in the current working directory.

If there is no compile_commands.json in the current working directory,
the default behavior of climbing up the parent folders of the source
file for which lsp is attached is retained.