fix: remove fudging of the system paths
The final correction of the system path so far set the /usr/local/
paths to the front again. That means that for example a local nvim
install takes precedence over a user-installed nvim in the HOME folder.

That is unexpected. Therefore, only keeping the setting of the DOTS/bin
folder and nothing else.