fix: remove fudging of the system paths (#94)
The final correction of the system path so far set the /usr/local/
paths to the front again. That means that for example a local nvim
install takes precedence over a user-installed nvim in the HOME folder.

That is unexpected. Hence, removing this special casing. Leaving the
file in place with a comment, though, in case that I need to remember
later how to tweak paths after they are set.