fix: fix searching with only quickfix window open (#120)
The removed autocmd group interfered with the scenario where only the
quickfix window was open. Since the definition of the autocommand group
was a little to indiscriminate, let's remove and replace it by a simple
invocation of :copen at the end of our search function.

Most importantly: This changes the behavior of an empty search. The
quickfix window will now no longer close. Instead, it will remain open
and be empty.

Part of #117