fix: no explicit homebrew update
Instead, rely on the 30 days cleanup
(HOMEBREW_CLEANUP_PERIODIC_FULL_DAYS) and on CI hope that someone else
updates from time to time as we disable the update of homebrew for CI
explicitly. For local runs, this remains unaffected.

This has all been done in the chase of a faster CI time for macOS
builds in #122.