fix(nvim): only source filetype.vim once (#279)
The previous gymnastics of calling `filetype off` and `filetype indent
plugin on` were only needed in Pathogen. I seem to have copied this
over when starting a fresh vimrc by using stevelosh's defaults.

With using Plug as the plugin manager instead, the filetype and syntax
enabling is handled already.

Removing the superfluous commands removes sourcing of ftoff and an
additional sourcing of filetype.vim. This saves about 10 ms of start up
time overall.