perf(nvim): do not use range context for git-log per default (#285)
To just browse through the history of a single file, the range given to
the Gclog command should not contain the entire file. This
unnecessarily slows down both the git command as well as the quickfix
parsing of the output: The entire file for each revision is parsed.

Instead, pass a zero context to the Gclog command.

This was erroneously introduced in 7c668d8.

Fixes #283.