fix(nvim): do not restore cursor position for git-log (#288)
For a simple git-log, restoring of the cursor will work for a single
file and single revision. As soon as I jump to the next file in the
quickfix list, the cursor position is gone.

If I want to restrict the git log to a specific set of lines, there is
a visual git log mapping that is superior. I will use that.

This commit removes the unnecessary wrapping of the git log Gclog
invocation in the RestoreCurPos.