fix(nvim): completion tests
The tests were broken after an update of nvim-cmp: Now most of the
functionality has to wait for the "next tick". That means that our
tests cannot wait and block the execution thread until the completion
is shown.

Instead, the tests now relinquish control when waiting and regain
control by setting a delayed timer.

The nvim-cmp-select-enter test was broken: It missed resetting of the
cursor for the final check of the contained text. This means that even
with the implementation was broken, the test would pass as it would not
find the incorrectly inserted test; for it was past that position
already.

Fixes #359

accept-perf: ci