feat!(script): automatically run bootstrap/upgrade for new changes
Prevent concurrent invocations when multiple shells are opened.
Expected flow:
- I open a new shell for the first time on a day.
- Updates to dotfiles are detected.
- I switch to a second new shell.
  -- pending update is detected.
  -- but I can start working on this shell.
  -- first shell continues updating.
- Dotfiles are updated from git.
- Bootstrap is run.
- Install is run.

This change also includes asking for the my sudo upfront and
changes the invocation for chsh to use sudo as well. This way I am
asked only once for the password in case there are changes and an
upgrade has to run.