chore(gdb): Remove non-actionable todo
There is no proper virtual package to install the correctly versioned
package for the debug symbols.

I also could not find a way to detect the currently installed version
easily to use that information to choose the right version of the
package for the debug symbols.

Since we have a test, this will fail if we for example update to
a later Ubuntu version.