fix(ci): only accept 4 sigma deviations
We initially accepted 6 sigma deviations as there were too few
measurements to have a more strict (and default) 4 sigma acceptance.

This is a flaw in git-perf and must be fixed there. I.e., if there are
not a minimum number of measurements, accept any measurement.

This change reverts 64307f8 and reintroduced the default 4 sigma
acceptance.