feat(script): user preset git user values when available
In codespaces, this information is already set and should be used (and
duplicated into the git credentials file. Without duplicating, we run
the risk that the original source of the preset values is altered. This
is a change in behavior.

This change allows dotfiles to be used in the setup of a codespace.

Part of #550.