fix(script): detect user/email for git from non-global context
codespaces expose the values in the system context, not in the global
context. To allow for both system and global context, do not narrow the
context. This also allows for local contexts to be elevated to global
context unfortunately. As there is no strong reason to prevent this
right now, let's keep it simple.

Part of #550.