refactor(script): convert test to bash script
Explicitly call test scripts instead of using their shebang.

This is a prep step for when the desired zsh is no longer guaranteed to
be at /bin/zsh, i.e., we want to make sure to test the final zsh and
not "some" zsh installed on a system.

topic:bash-test