feat(zsh): load zsh with exec in profile instead of chsh
As prep for using a nix-installed zsh, do not directly change the shell
for the user. In case that the nix-installed shell is not available or
fails to start, we risk a complete lock out.

With a safe-exec of the zsh binary in the profile, we can fallback to
the default login shell of the user.

topic: safe-zsh
relative-topic: faster-ci