fix: no login shell for tests (#941)
For #939, we cannot use a login shell as the nix daemon script putting
the profile bin directory in front of everything else in the PATH will
already have run once and exported the env var to prevent duplicated
running. Running as a login shell for tests will only run the macOS
specific helpers to put system binaries in front of the PATH again. This
ultimately will mean that for a nested login shell, the system path
libraries will precede the nix installed ones.